ITS ALL DIFFERENT:
Organized the file structure and updated imports
Added some logic for DB connection in utils.py and bulk_inster.py - this still needs some further isolation
