+++ app.py
@@ -1,6 +1,6 @@
 import os
 from dotenv import load_dotenv
-from models import db, connect_db
+from database_utils.models import db, connect_db
 
 from flask import Flask
 DEV = True
@@ -9,7 +9,6 @@ load_dotenv()
 app = Flask(__name__)
 
 database_url = os.environ['TEST_DATABASE_URL'] if DEV else os.environ["DATABASE_URL"]
-database_url = database_url.replace('postgresql://', 'postgres://') if DEV else database_url.replace('postgres://', 'postgresql://')
 app.config['SQLALCHEMY_ECHO'] = False
 app.config['DEBUG_TB_INTERCEPT_REDIRECTS'] = False
 app.config['SECRET_KEY'] = os.environ['SECRET_KEY']
